+++ src/pytorch_lightning/utilities/seed.py
@@ -13,13 +13,9 @@
 # limitations under the License.
 """Utilities to help with reproducibility of models."""
 from contextlib import contextmanager
-from typing import Any, Generator
+from typing import Generator
 
 from lightning_fabric.utilities.seed import _collect_rng_states, _set_rng_states
-from lightning_fabric.utilities.seed import pl_worker_init_function as new_pl_worker_init_function
-from lightning_fabric.utilities.seed import reset_seed as new_reset_seed
-from lightning_fabric.utilities.seed import seed_everything as new_seed_everything
-from pytorch_lightning.utilities.rank_zero import rank_zero_deprecation
 
 
 @contextmanager
@@ -46,27 +42,3 @@ def isolate_rng(include_cuda: bool = True) -> Generator[None, None, None]:
     states = _collect_rng_states(include_cuda)
     yield
     _set_rng_states(states)
-
-
-def seed_everything(*args: Any, **kwargs: Any) -> Any:
-    rank_zero_deprecation(
-        "`pytorch_lightning.utilities.seed.seed_everything` has been deprecated in v1.8.0 and will be"
-        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.seed_everything` instead."
-    )
-    return new_seed_everything(*args, **kwargs)
-
-
-def reset_seed() -> None:
-    rank_zero_deprecation(
-        "`pytorch_lightning.utilities.seed.reset_seed` has been deprecated in v1.8.0 and will be"
-        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.reset_seed` instead."
-    )
-    return new_reset_seed()
-
-
-def pl_worker_init_function(*args: Any, **kwargs: Any) -> None:
-    rank_zero_deprecation(
-        "`pytorch_lightning.utilities.seed.pl_worker_init_function` has been deprecated in v1.8.0 and will be"
-        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.pl_worker_init_function` instead."
-    )
-    return new_pl_worker_init_function(*args, **kwargs)